Extract knob-cartesian enumeration to _enumeration.py sibling module
010_partition_loops.py carried ~350 lines of "which (BN, BM, FM, FN, BK,
SPLITK, BR) tuples are admissible, in what order" logic — pure-function
enumeration with no IR manipulation. Moved to a non-pass sibling module
so the planner file is focused on IR + planning + materialization.

Moved to `_enumeration.py`:
- Knob globals (BN, BM, FM, FN, BK, SPLITK, BR) + planner-knob tuple
- Choice constants (_TUNE_AXIS_CHOICES, _BK_CANDIDATES, _SPLITK_CANDIDATES,
  _BR_CANDIDATES, _TUNE_F_CHOICES, _MAX_CELLS_PER_THREAD)
- `TileParams` dataclass (the cartesian product element)
- `planner_pin_set()` (env-pin probe)
- Per-mode priority functions (_priority_matmul / _priority_pointwise /
  _priority_reduce)
- `enumerate_cartesian()` (mode-dispatched wrapper) +
  `_enumerate_cartesian_impl()` (pure cartesian generator)
- `_divisors_up_to()` helper

Public API: `TileParams`, `enumerate_cartesian`, `planner_pin_set`, and
the 7 Knob globals. `_enumerate_cartesian_impl` + helpers stay private.

`010_partition_loops.py` keeps all the IR-manipulating logic:
- Plan / _Plan / KernelShape / Role / _BuildSkipped
- rewrite() + _plan_kernel + _materialize
- Body builders (_build_split_body, _build_register_blocked_body,
  _replace_k_loops, _build_k_sigma, _wrap_tower, etc.)
- Kernel classification (_outer_free_loop_chain,
  _classify_fused_prologue, _contains_matmul_reduce)

Re-imports the knob globals from `_enumeration.py` so the planner's
`rewrite()` Level definitions and `_materialize` knob-stamping work as
before. Tests access `_planner.BR` etc. through the planner's namespace —
those resolve to the re-imported names, no test changes needed.

Verification: make test (1377 passed, 27 skipped); make lint clean.
Net: partition_loops.py 1507 -> 1163 lines (-344, -23%).

After three extractions (fork-tree, splitk-gate, enumeration), the
planner module is down from 1710 to 1163 lines (-32%); the structural
rewrites, knob-cartesian enumeration, and Fork-tree builder each live
in their own focused files.
